+++ Moobeez/Classes/Gogz/External SDKs/SBJson/SBJsonStreamParser.m
@@ -135,7 +135,7 @@ static NSNull *kNull;
 }
 
 - (void)maxDepthError {
-    self.error = [NSString stringWithFormat:@"Input depth exceeds max depth of %u", maxDepth];
+    self.error = [NSString stringWithFormat:@"Input depth exceeds max depth of %lu", (unsigned long)maxDepth];
     self.state = [SBJsonStreamParserStateError sharedInstance];
 }
